--- cypress/integration/Log/EndlessLoopX.js
@@ -1,1 +1,0 @@
-require('../../support/TestListManager').AddToCypressTestList('Log.EndlessLoopX')
