+++ cypress/integration/Log/WhatsYourName.js
@@ -1,1 +1,23 @@
-require('../../support/TestListManager').AddToCypressTestList('Log.WhatsYourName')
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.  
+ * Licensed under the MIT License.
+*/
+
+const models = require('../../support/Models')
+const modelPage = require('../../support/components/ModelPage')
+const logDialogsGrid = require('../../support/components/LogDialogsGrid')
+const logDialogModal = require('../../support/components/LogDialogModal')
+const common = require('../../support/Common')
+
+export function WhatsYourName()
+{
+  models.ImportModel('z-logMyName', 'z-nameTrained.cl')
+
+  modelPage.NavigateToLogDialogs()
+  cy.WaitForTrainingStatusCompleted()
+  logDialogsGrid.CreateNewLogDialogButton()
+
+  logDialogModal.TypeYourMessageValidateResponse("Hello", common.whatsYourName)
+
+  logDialogModal.ClickDoneTestingButton()
+}
